Solve Day 2 problems
I extracted a simple method to read data from a file which also trims and filters empty lines. Parsing the non-empty lines into meaningful input will be the responsibility of challenge-specific code.

Parsing the commands for this challenge was done with funcj-parser, which is overkill now but in future challenges I assume that the increased power of the library will be helpful.