Solve the Day 14 challenges
Today took some thought because the naive solution, which is to actually create the expanded string, consumes (at least) hundreds of gigabytes of memory. The key detail is that nothing in the problem statement requires us to know _where_ in the polymer each pair occurs, so we can compute everything we need using a histogram of the pairs instead of the actual string.