Solve the Day 16 challenges
This one was a doozy. Given a protocol buffer-like encoding, parse and evaluate mathematical expressions. I've learned that both Java and Kotlin are _terrible_ at dealing with bits (as opposed to bytes). BitSet seems like it's supposed to be the random access collection for bits, but gaps in its API made it nigh unusable for this challenge.